Changes to be committed: modified:   frontend/src/components/header/index.jsx renamed:    frontend/src/components/header/mobile-sidebar.jsx -> frontend/src/components/header/mobileSidebar.jsx renamed:    frontend/src/components/header/nav-links.jsx -> frontend/src/components/header/navLinks.jsx renamed:    frontend/src/components/header/search-input.jsx -> frontend/src/components/header/searchInput.jsx renamed:    frontend/src/components/header/theme-toggler.jsx -> frontend/src/components/header/themeToggler.jsx renamed:    frontend/src/components/loading-button.jsx -> frontend/src/components/loadingButton.jsx renamed:    frontend/src/components/password-input.jsx -> frontend/src/components/passwordInput.jsx renamed:    frontend/src/components/user-button.jsx -> frontend/src/components/userButton.jsx modified:   frontend/src/pages/auth/login.jsx new file:   frontend/src/pages/blogs.jsx
