+++ frontend/src/components/header/navLinks.jsx
@@ -41,4 +41,4 @@ const NavLinks = () => {
   );
 };
 
-export default NavLinks;
+export default NavLinks;
