+++ frontend/src/components/userButton.jsx
@@ -109,4 +109,4 @@ export const UserButton = () => {
       </PopoverContent>
     </Popover>
   );
-};
+};
