+++ frontend/src/pages/blogs.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Blogs = () => {
+  return (
+    <div>
+      Blog Page
+    </div>
+  )
+}
+
+export default Blogs
